// src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export interface BaseUserReadable {
  firstName?: string;
  lastName?: string;
  readonly email?: string;
  nickname?: string;
  gravatarEmailAddress?: string;
  picture?: string;
  avatarImageSource?: AvatarImageSource;
}

export interface BaseUserWritable {
  firstName?: string;
  lastName?: string;
  nickname?: string;
  gravatarEmailAddress?: string;
  picture?: string;
  avatarImageSource?: AvatarImageSource;
}

export type UserInfoReadable = BaseUserReadable;

export type UserInfoWritable = BaseUserWritable;

export type PatchUserInfoReadable = BaseUserReadable;

export type PatchUserInfoWritable = BaseUserWritable;

export type AvatarImageSource = 'GRAVATAR' | 'MANUAL';

export interface _Error {
  code: number;
  message: string;
}

export interface GetUserInfoData {
  body?: never;
  path?: never;
  query?: never;
  url: '/user-info';
}

export interface GetUserInfoErrors {
  /**
   * Unexpected error
   */
  default: _Error;
}

export type GetUserInfoError = GetUserInfoErrors[keyof GetUserInfoErrors];

export interface GetUserInfoResponses {
  /**
   * Successful operation
   */
  200: UserInfoReadable;
}

export type GetUserInfoResponse = GetUserInfoResponses[keyof GetUserInfoResponses];

export interface UpdateUserInfoData {
  /**
   * Partial details of user info to update
   */
  body: PatchUserInfoWritable;
  path?: never;
  query?: never;
  url: '/user-info';
}

export interface UpdateUserInfoErrors {
  /**
   * Unexpected error
   */
  default: _Error;
}

export type UpdateUserInfoError = UpdateUserInfoErrors[keyof UpdateUserInfoErrors];

export interface UpdateUserInfoResponses {
  /**
   * Successful operation
   */
  200: UserInfoReadable;
}

export type UpdateUserInfoResponse = UpdateUserInfoResponses[keyof UpdateUserInfoResponses];

export interface ClientOptions {
  baseUrl: 'https://user-service.dajohnston.co.uk/api/v1' | 'http://localhost:3000/api/v1' | (string & {});
}